Refactor Hue Driver ahead of adding new devices.
