+++ drivers/SmartThings/philips-hue/src/consts.lua
@@ -1,0 +1,9 @@
+local Consts = {}
+
+--- minimum colortemp value from Hue
+Consts.DEFAULT_MIREK = 153
+Consts.MIN_TEMP_KELVIN_COLOR_AMBIANCE = 2000
+Consts.MIN_TEMP_KELVIN_WHITE_AMBIANCE = 2200
+Consts.MAX_TEMP_KELVIN = 6500
+
+return Consts
