+++ drivers/SmartThings/philips-hue/src/disco/init.lua
@@ -1,4 +1,4 @@
-local log = require "log"
+local log = require "logjam"
 local socket = require "cosock.socket"
 
 local mdns = require "st.mdns"
@@ -7,118 +7,75 @@ local st_utils = require "st.utils"
 
 local Fields = require "fields"
 local HueApi = require "hue.api"
+local HueDeviceTypes = require "hue_device_types"
+
 local utils = require "utils"
 
 local SERVICE_TYPE = "_hue._tcp"
 local DOMAIN = "local"
 
+---@class DiscoveredChildDeviceHandler
+---@field public handle_discovered_device fun(driver: HueDriver, bridge_id: string, api_instance: PhilipsHueApi, resource_id: string, device_service_info: table, device_state_disco_cache: table<string, table>, st_metadata_callback: fun(driver: HueDriver, metadata: table)?)
+
 -- This `api_keys` table is an in-memory fall-back table. It gets overwritten
 -- with a reference to a driver datastore table before the Driver's `run` loop
 -- can get spun up in `init.lua`.
+---@class HueDiscovery
+---@field public api_keys table<string,string>
+---@field public disco_api_instances table<string,PhilipsHueApi>
+---@field public device_state_disco_cache table<string,table<string,any>>
 local HueDiscovery = {
   api_keys = {},
   disco_api_instances = {},
-  light_state_disco_cache = {},
+  device_state_disco_cache = {},
   ServiceType = SERVICE_TYPE,
   Domain = DOMAIN,
   discovery_active = false
 }
 
-local supported_resource_types = {
-  light = true
-}
-
--- "forward declarations"
-local discovered_bridge_callback, discovered_device_callback, is_device_service_supported,
-process_discovered_light
+-- Lazy-load the discovered device handlers so we only load the code we need
+---@type table<string,DiscoveredChildDeviceHandler>
+local discovered_device_handlers = utils.lazy_handler_loader("disco")
 
 ---comment
+---@param svc_info HueServiceInfo
+---@return boolean
+local function is_device_service_supported(svc_info)
+  return discovered_device_handlers[svc_info.rtype or ""] ~= nil
+end
+
 ---@param driver HueDriver
----@param _ table
----@param should_continue function
-function HueDiscovery.discover(driver, _, should_continue)
-  if HueDiscovery.discovery_active then
-    log.info("Hue discovery already in progress, ignoring new discovery request")
+---@param bridge_id string
+---@param svc_info HueServiceInfo
+---@param device_info table
+local function discovered_device_callback(driver, bridge_id, svc_info, device_info)
+  local v1_dni = bridge_id .. "/" .. (device_info.id_v1 or "UNKNOWN"):gsub("/lights/", "")
+  local v2_resource_id = svc_info.rid or ""
+  if driver:get_device_by_dni(v1_dni) or driver.hue_identifier_to_device_record[v2_resource_id] then return end
+
+  local api_instance = HueDiscovery.disco_api_instances[bridge_id]
+  if not api_instance then
+    log.warn("No API instance for bridge_id ", bridge_id)
     return
   end
 
-  log.info_with({ hub_logs = true }, "Starting Hue discovery")
-  HueDiscovery.discovery_active = true
-
-  while should_continue() do
-    local known_identifier_to_device_map = {}
-    local computed_mac_addresses = {}
-    for _, device in ipairs(driver:get_devices()) do
-      -- the bridge won't have a parent assigned key so we give that boolean short circuit preference
-      local dni = device.parent_assigned_child_key or device.device_network_id
-      known_identifier_to_device_map[dni] = device
-      local ipv4 = device:get_field(Fields.IPV4);
-      if ipv4 then
-        computed_mac_addresses[ipv4] = dni
-      end
-    end
-
-    HueDiscovery.do_mdns_scan(driver)
-    HueDiscovery.search_for_bridges(driver, computed_mac_addresses, function(hue_driver, bridge_ip, bridge_id)
-      discovered_bridge_callback(hue_driver, bridge_ip, bridge_id, known_identifier_to_device_map)
-    end)
-    socket.sleep(1.0)
-  end
-  HueDiscovery.discovery_active = false
-  log.info_with({ hub_logs = true }, "Ending Hue discovery")
-end
-
----comment
----@param driver HueDriver
----@param computed_mac_addresses table<string,string>
----@param callback fun(driver: HueDriver, ip: string, id: string)
-function HueDiscovery.search_for_bridges(driver, computed_mac_addresses, callback)
-  local scanned_bridges = driver.datastore.bridge_netinfo or {}
-  for bridge_id, bridge_info in pairs(scanned_bridges) do
-    if type(callback) == "function" and bridge_info ~= nil then
-      callback(driver, bridge_info.ip, bridge_id)
-    else
-      log.warn(
-        "Argument passed in `callback` position for "
-        .. "`HueDiscovery.search_for_bridges` is not a function"
-      )
-    end
-  end
-end
-
-function HueDiscovery.scan_bridge_and_update_devices(driver, bridge_id)
-  if driver.ignored_bridges[bridge_id] then return end
-
-  local known_identifier_to_device_map = {}
-  for _, device in ipairs(driver:get_devices()) do
-    -- the bridge won't have a parent assigned key so we give that boolean short circuit preference
-    local dni = device.parent_assigned_child_key or device.device_network_id
-    known_identifier_to_device_map[dni] = device
-  end
-
-  local known_bridge_device = known_identifier_to_device_map[bridge_id]
-  if known_bridge_device and known_bridge_device:get_field(HueApi.APPLICATION_KEY_HEADER) then
-    HueDiscovery.api_keys[bridge_id] = known_bridge_device:get_field(HueApi.APPLICATION_KEY_HEADER)
-  end
-
-  HueDiscovery.search_bridge_for_supported_devices(driver, bridge_id, HueDiscovery.disco_api_instances[bridge_id],
-    function(hue_driver, svc_info, device_info)
-      discovered_device_callback(hue_driver, bridge_id, svc_info, device_info, known_identifier_to_device_map)
-    end,
-    "[Discovery: " ..
-    (known_bridge_device.label or bridge_id or known_bridge_device.id or "unknown bridge") .. " bridge re-scan]",
-    true
+  HueDiscovery.handle_discovered_child_device(
+    driver,
+    bridge_id,
+    api_instance,
+    svc_info,
+    device_info
   )
 end
 
+-- "forward declarations"
 ---@param driver HueDriver
 ---@param bridge_ip string
 ---@param bridge_id string
----@param known_identifier_to_device_map table<string,HueDevice>
-discovered_bridge_callback = function(driver, bridge_ip, bridge_id, known_identifier_to_device_map)
+local function discovered_bridge_callback(driver, bridge_ip, bridge_id)
   if driver.ignored_bridges[bridge_id] then return end
 
-  local known_bridge_device = known_identifier_to_device_map[bridge_id]
+  local known_bridge_device = driver:get_device_by_dni(bridge_id)
   if known_bridge_device and known_bridge_device:get_field(HueApi.APPLICATION_KEY_HEADER) then
     HueDiscovery.api_keys[bridge_id] = known_bridge_device:get_field(HueApi.APPLICATION_KEY_HEADER)
   end
@@ -126,7 +83,6 @@ discovered_bridge_callback = function(driver, bridge_ip, bridge_id, known_identi
   if known_bridge_device ~= nil
       and driver.joined_bridges[bridge_id]
       and HueDiscovery.api_keys[bridge_id]
-      and known_bridge_device:get_field(Fields._INIT)
   then
     log.info_with({ hub_logs = true }, string.format("Scanning bridge %s for devices...", bridge_id))
 
@@ -137,12 +93,16 @@ discovered_bridge_callback = function(driver, bridge_ip, bridge_id, known_identi
           utils.labeled_socket_builder((known_bridge_device.label or bridge_id or known_bridge_device.id or "unknown bridge"))
         )
 
-    HueDiscovery.search_bridge_for_supported_devices(driver, bridge_id, HueDiscovery.disco_api_instances[bridge_id],
+    HueDiscovery.search_bridge_for_supported_devices(
+      driver,
+      bridge_id,
+      HueDiscovery.disco_api_instances[bridge_id],
       function(hue_driver, svc_info, device_info)
-        discovered_device_callback(hue_driver, bridge_id, svc_info, device_info, known_identifier_to_device_map)
+        discovered_device_callback(hue_driver, bridge_id, svc_info, device_info)
       end,
       "[Discovery: " ..
-      (known_bridge_device.label or bridge_id or known_bridge_device.id or "unknown bridge") .. " bridge scan]"
+      (known_bridge_device.label or bridge_id or known_bridge_device.id or "unknown bridge") ..
+      " bridge scan]"
     )
     return
   end
@@ -209,10 +169,81 @@ discovered_bridge_callback = function(driver, bridge_ip, bridge_id, known_identi
   end
 end
 
+---comment
+---@param driver HueDriver
+---@param _ table
+---@param should_continue function
+function HueDiscovery.discover(driver, _, should_continue)
+  if HueDiscovery.discovery_active then
+    log.info("Hue discovery already in progress, ignoring new discovery request")
+    return
+  end
+
+  log.info_with({ hub_logs = true }, "Starting Hue discovery")
+  HueDiscovery.discovery_active = true
+
+  while should_continue() do
+    HueDiscovery.do_mdns_scan(driver)
+    HueDiscovery.search_for_bridges(
+      driver,
+      function(hue_driver, bridge_ip, bridge_id)
+        discovered_bridge_callback(hue_driver, bridge_ip, bridge_id)
+      end
+    )
+    socket.sleep(1.0)
+  end
+  HueDiscovery.discovery_active = false
+  log.info_with({ hub_logs = true }, "Ending Hue discovery")
+end
+
+---@param driver HueDriver
+---@param callback fun(driver: HueDriver, ip: string, id: string)
+function HueDiscovery.search_for_bridges(driver, callback)
+  local scanned_bridges = driver.datastore.bridge_netinfo or {}
+  for bridge_id, bridge_info in pairs(scanned_bridges) do
+    if type(callback) == "function" and bridge_info ~= nil then
+      callback(driver, bridge_info.ip, bridge_id)
+    else
+      log.warn(
+        "Argument passed in `callback` position for "
+        .. "`HueDiscovery.search_for_bridges` is not a function"
+      )
+    end
+  end
+end
+
+---@param driver HueDriver
+---@param bridge_id string
+function HueDiscovery.scan_bridge_and_update_devices(driver, bridge_id)
+  if driver.ignored_bridges[bridge_id] then return end
+
+  local known_bridge_device = driver:get_device_by_dni(bridge_id)
+  if known_bridge_device then
+    if known_bridge_device:get_field(HueApi.APPLICATION_KEY_HEADER) then
+      HueDiscovery.api_keys[bridge_id] = known_bridge_device:get_field(HueApi.APPLICATION_KEY_HEADER)
+    end
+
+    HueDiscovery.search_bridge_for_supported_devices(
+      driver,
+      bridge_id,
+      HueDiscovery.disco_api_instances[bridge_id],
+      function(hue_driver, svc_info, device_info)
+        discovered_device_callback(hue_driver, bridge_id, svc_info, device_info)
+      end,
+      "[Discovery: " ..
+      (known_bridge_device.label or bridge_id or known_bridge_device.id or "unknown bridge") ..
+      " bridge re-scan]",
+      true
+    )
+  end
+end
+
 ---@param driver HueDriver
 ---@param bridge_id string
 ---@param api_instance PhilipsHueApi
----@param callback fun(driver: HueDriver, svc_info: table, device_data: table)
+---@param callback fun(driver: HueDriver, svc_info: HueServiceInfo, device_data: table)
+---@param log_prefix string?
+---@param do_delete boolean?
 function HueDiscovery.search_bridge_for_supported_devices(driver, bridge_id, api_instance, callback, log_prefix, do_delete)
   local prefix = ""
   if type(log_prefix) == "string" and #log_prefix > 0 then prefix = log_prefix .. " " end
@@ -234,28 +265,36 @@ function HueDiscovery.search_bridge_for_supported_devices(driver, bridge_id, api
 
   local device_is_joined_to_bridge = {}
   for _, device_data in ipairs(devices.data or {}) do
+    local primary_device_service
     for _, svc_info in ipairs(device_data.services or {}) do
       if is_device_service_supported(svc_info) then
-        device_is_joined_to_bridge[device_data.id] = true
-        log.info_with({ hub_logs = true }, string.format(
-          prefix ..
-          "Processing supported svc [rid: %s | type: %s] for Hue device [v2_id: %s | v1_id: %s], with Hue provided name: %s",
-          svc_info.rid, svc_info.rtype, device_data.id, device_data.id_v1, device_data.metadata.name
-        ))
-        if type(callback) == "function" then
-          callback(driver, svc_info, device_data)
-        else
-          log.warn(
-            prefix .. "Argument passed in `callback` position for "
-            .. "`HueDiscovery.search_bridge_for_supported_devices` is not a function"
-          )
+        driver.services_for_device_rid[device_data.id] = driver.services_for_device_rid[device_data.id] or {}
+        driver.services_for_device_rid[device_data.id][svc_info.rid] = svc_info.rtype
+        if HueDeviceTypes.can_join_device_for_service(svc_info.rtype) then
+          primary_device_service = svc_info
+          device_is_joined_to_bridge[device_data.id] = true
         end
       end
+    end
+    if primary_device_service and type(callback) == "function" then
+      log.info_with(
+        { hub_logs = true }, string.format(
+        prefix ..
+        "Processing supported svc [rid: %s | type: %s] for Hue device [v2_id: %s | v1_id: %s], with Hue provided name: %s",
+        primary_device_service.rid, primary_device_service.rtype, device_data.id, device_data.id_v1, device_data.metadata.name
+        )
+      )
+      callback(driver, primary_device_service, device_data)
+    else
+      log.warn(
+        prefix .. "Argument passed in `callback` position for "
+        .. "`HueDiscovery.search_bridge_for_supported_devices` is not a function"
+      )
     end
   end
 
   if do_delete then
-    for _, device in ipairs(driver:get_devices()) do ---@cast device HueDevice
+    for _, device in ipairs(driver:get_devices()) do
       -- We're only interested in processing child/non-bridge devices here.
       if utils.is_bridge(driver, device) then goto continue end
       local not_known_to_bridge = device_is_joined_to_bridge[device:get_field(Fields.HUE_DEVICE_ID) or ""]
@@ -264,7 +303,7 @@ function HueDiscovery.search_bridge_for_supported_devices(driver, bridge_id, api
       local is_child_of_bridge = parent_bridge_device and (parent_bridge_device:get_field(Fields.BRIDGE_ID) == bridge_id)
       if parent_bridge_device and is_child_of_bridge and not not_known_to_bridge and device.id then
         device.log.info(string.format("Device is no longer joined to Hue Bridge %q, deleting", parent_bridge_device.label))
-        driver:do_hue_light_delete(device)
+        driver:do_hue_child_delete(device)
       end
       ::continue::
     end
@@ -273,107 +312,22 @@ end
 
 ---@param driver HueDriver
 ---@param bridge_id string
+---@param api_instance PhilipsHueApi
 ---@param svc_info table
 ---@param device_info table
----@param known_identifier_to_device_map table<string,HueDevice>
-discovered_device_callback = function(driver, bridge_id, svc_info, device_info, known_identifier_to_device_map)
-  local v1_dni = bridge_id .. "/" .. (device_info.id_v1 or "UNKNOWN"):gsub("/lights/", "")
-  local edge_dni = svc_info.rid or ""
-  if known_identifier_to_device_map[v1_dni] or known_identifier_to_device_map[edge_dni] then return end
-  if svc_info.rtype == "light" then
-    process_discovered_light(driver, bridge_id, svc_info.rid, device_info, known_identifier_to_device_map)
-  end
+function HueDiscovery.handle_discovered_child_device(driver, bridge_id, api_instance, svc_info, device_info)
+  discovered_device_handlers[svc_info.rtype].handle_discovered_device(
+    driver,
+    bridge_id,
+    api_instance,
+    svc_info.rid,
+    device_info,
+    HueDiscovery.device_state_disco_cache,
+    driver.try_create_device
+  )
 end
 
 ---@param driver HueDriver
----@param bridge_id string
----@param resource_id string
----@param device_info table
----@param known_identifier_to_device_map table<string,HueDevice>
-process_discovered_light = function(driver, bridge_id, resource_id, device_info, known_identifier_to_device_map)
-  local api_instance = HueDiscovery.disco_api_instances[bridge_id]
-  if not api_instance then
-    log.warn("No API instance for bridge_id ", bridge_id)
-    return
-  end
-
-  local light_resource, err, _ = api_instance:get_light_by_id(resource_id)
-  if err ~= nil or not light_resource then
-    log.error_with({ hub_logs = true },
-      string.format("Error getting light info for %s: %s", device_info.product_data.product_name,
-        (err or "unexpected nil in error position")))
-    return
-  end
-
-  if light_resource.errors and #light_resource.errors > 0 then
-    log.error_with({ hub_logs = true }, "Errors found in API response:")
-    for idx, err in ipairs(light_resource.errors) do
-      log.error_with({ hub_logs = true }, st_utils.stringify_table(err, "Error " .. idx, true))
-    end
-    return
-  end
-
-  for _, light in ipairs(light_resource.data or {}) do
-    local profile_ref
-
-    if light.color then
-      if light.color_temperature then
-        profile_ref = "white-and-color-ambiance"
-      else
-        profile_ref = "legacy-color"
-      end
-    elseif light.color_temperature then
-      profile_ref = "white-ambiance" -- all color temp products support `white` (dimming)
-    elseif light.dimming then
-      profile_ref = "white"          -- `white` refers to dimmable and includes filament bulbs
-    else
-      log.warn(
-        string.format(
-          "Light resource [%s] does not seem to be A White/White-Ambiance/White-Color-Ambiance device, currently unsupported"
-          ,
-          resource_id
-        )
-      )
-      goto continue
-    end
-
-    local bridge_device = known_identifier_to_device_map[bridge_id]
-
-    local create_device_msg = {
-      type = "EDGE_CHILD",
-      label = light.metadata.name,
-      vendor_provided_label = device_info.product_data.product_name,
-      profile = profile_ref,
-      manufacturer = device_info.product_data.manufacturer_name,
-      model = device_info.product_data.model_id,
-      parent_device_id = bridge_device.id,
-      parent_assigned_child_key = light.id,
-    }
-
-    HueDiscovery.light_state_disco_cache[light.id] = {
-      hue_provided_name = light.metadata.name,
-      id = light.id,
-      on = light.on,
-      color = light.color,
-      dimming = light.dimming,
-      color_temperature = light.color_temperature,
-      mode = light.mode,
-      parent_device_id = bridge_device.id,
-      hue_device_id = light.owner.rid,
-      hue_device_data = device_info
-    }
-
-    driver:try_create_device(create_device_msg)
-    -- rate limit ourself.
-    socket.sleep(0.1)
-    ::continue::
-  end
-end
-
-is_device_service_supported = function(svc_info)
-  return supported_resource_types[svc_info.rtype or ""]
-end
-
 function HueDiscovery.do_mdns_scan(driver)
   local bridge_netinfo = driver.datastore.bridge_netinfo
   local mdns_responses, err = mdns.discover(HueDiscovery.ServiceType, HueDiscovery.Domain)
@@ -445,6 +399,7 @@ function HueDiscovery.do_mdns_scan(driver)
       end
     end
 
+    ---@type boolean?
     local update_needed = false
     if not utils.deep_table_eq((bridge_netinfo[bridge_id] or {}), bridge_info) then
       bridge_netinfo[bridge_id] = bridge_info
